+++ modules/ntshengn_system_module_interface.h
@@ -19,7 +19,7 @@ namespace NtshEngn {
 		virtual ~SystemModuleInterface() {}
 
 		virtual void init() = 0;
-		virtual void update(double dt) = 0;
+		virtual void update(float dt) = 0;
 		virtual void destroy() = 0;
 
 		virtual const ComponentMask getComponentMask() const {
